pyproject.toml
[project]
name = "bruinbuddy"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.50.0,<2.0.0)",
    "strands-agents (>=1.10.0,<2.0.0)",
    "boto3 (>=1.40.45,<2.0.0)",
    "awscli (>=1.42.45,<2.0.0)"
]

[tool.poetry]
packages = [{include = "bruinbuddy", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
